refactor: move rpc.rs to node/rest_api
- this a first step to integrate REST API into the node
