+++ src/setup/node/rest_api/mod.rs
@@ -1,0 +1,8 @@
+//! A REST API implementation is named RPC in the go-algorand code base.
+//!
+//! There are two REST API versions for algod:
+//! - [V1](https://developer.algorand.org/docs/rest-apis/algod/v1/) - which is deprecated but still used by the node.
+//! - [V2](https://developer.algorand.org/docs/rest-apis/algod/v2/)
+
+pub mod client;
+pub mod message;
